+++ sdk/dotnet/Rds/Instance.cs
@@ -106,7 +106,7 @@ namespace Pulumi.AliCloud.Rds
         public Output<string> DbInstanceStorageType { get; private set; } = null!;
 
         /// <summary>
-        /// The key id of the KMS. Used for encrypting a disk if not null. Only for PostgreSQL.
+        /// The key id of the KMS. Used for encrypting a disk if not null. Only for PostgreSQL, MySQL.
         /// </summary>
         [Output("encryptionKey")]
         public Output<string?> EncryptionKey { get; private set; } = null!;
@@ -361,7 +361,7 @@ namespace Pulumi.AliCloud.Rds
         public Input<string>? DbInstanceStorageType { get; set; }
 
         /// <summary>
-        /// The key id of the KMS. Used for encrypting a disk if not null. Only for PostgreSQL.
+        /// The key id of the KMS. Used for encrypting a disk if not null. Only for PostgreSQL, MySQL.
         /// </summary>
         [Input("encryptionKey")]
         public Input<string>? EncryptionKey { get; set; }
@@ -595,7 +595,7 @@ namespace Pulumi.AliCloud.Rds
         public Input<string>? DbInstanceStorageType { get; set; }
 
         /// <summary>
-        /// The key id of the KMS. Used for encrypting a disk if not null. Only for PostgreSQL.
+        /// The key id of the KMS. Used for encrypting a disk if not null. Only for PostgreSQL, MySQL.
         /// </summary>
         [Input("encryptionKey")]
         public Input<string>? EncryptionKey { get; set; }