+++ sdk/go/alicloud/rds/instance.go
@@ -106,7 +106,7 @@ type Instance struct {
 	// - cloud_essd2: specifies to use enhanced SSDs (ESSDs).
 	// - cloud_essd3: specifies to use enhanced SSDs (ESSDs).
 	DbInstanceStorageType pulumi.StringOutput `pulumi:"dbInstanceStorageType"`
-	// The key id of the KMS. Used for encrypting a disk if not null. Only for PostgreSQL.
+	// The key id of the KMS. Used for encrypting a disk if not null. Only for PostgreSQL, MySQL.
 	EncryptionKey pulumi.StringPtrOutput `pulumi:"encryptionKey"`
 	// Database type. Value options: MySQL, SQLServer, PostgreSQL, and PPAS.
 	Engine pulumi.StringOutput `pulumi:"engine"`
@@ -233,7 +233,7 @@ type instanceState struct {
 	// - cloud_essd2: specifies to use enhanced SSDs (ESSDs).
 	// - cloud_essd3: specifies to use enhanced SSDs (ESSDs).
 	DbInstanceStorageType *string `pulumi:"dbInstanceStorageType"`
-	// The key id of the KMS. Used for encrypting a disk if not null. Only for PostgreSQL.
+	// The key id of the KMS. Used for encrypting a disk if not null. Only for PostgreSQL, MySQL.
 	EncryptionKey *string `pulumi:"encryptionKey"`
 	// Database type. Value options: MySQL, SQLServer, PostgreSQL, and PPAS.
 	Engine *string `pulumi:"engine"`
@@ -321,7 +321,7 @@ type InstanceState struct {
 	// - cloud_essd2: specifies to use enhanced SSDs (ESSDs).
 	// - cloud_essd3: specifies to use enhanced SSDs (ESSDs).
 	DbInstanceStorageType pulumi.StringPtrInput
-	// The key id of the KMS. Used for encrypting a disk if not null. Only for PostgreSQL.
+	// The key id of the KMS. Used for encrypting a disk if not null. Only for PostgreSQL, MySQL.
 	EncryptionKey pulumi.StringPtrInput
 	// Database type. Value options: MySQL, SQLServer, PostgreSQL, and PPAS.
 	Engine pulumi.StringPtrInput
@@ -411,7 +411,7 @@ type instanceArgs struct {
 	// - cloud_essd2: specifies to use enhanced SSDs (ESSDs).
 	// - cloud_essd3: specifies to use enhanced SSDs (ESSDs).
 	DbInstanceStorageType *string `pulumi:"dbInstanceStorageType"`
-	// The key id of the KMS. Used for encrypting a disk if not null. Only for PostgreSQL.
+	// The key id of the KMS. Used for encrypting a disk if not null. Only for PostgreSQL, MySQL.
 	EncryptionKey *string `pulumi:"encryptionKey"`
 	// Database type. Value options: MySQL, SQLServer, PostgreSQL, and PPAS.
 	Engine string `pulumi:"engine"`
@@ -494,7 +494,7 @@ type InstanceArgs struct {
 	// - cloud_essd2: specifies to use enhanced SSDs (ESSDs).
 	// - cloud_essd3: specifies to use enhanced SSDs (ESSDs).
 	DbInstanceStorageType pulumi.StringPtrInput
-	// The key id of the KMS. Used for encrypting a disk if not null. Only for PostgreSQL.
+	// The key id of the KMS. Used for encrypting a disk if not null. Only for PostgreSQL, MySQL.
 	EncryptionKey pulumi.StringPtrInput
 	// Database type. Value options: MySQL, SQLServer, PostgreSQL, and PPAS.
 	Engine pulumi.StringInput