+++ sdk/nodejs/rds/instance.ts
@@ -141,7 +141,7 @@ export class Instance extends pulumi.CustomResource {
      */
     public readonly dbInstanceStorageType!: pulumi.Output<string>;
     /**
-     * The key id of the KMS. Used for encrypting a disk if not null. Only for PostgreSQL.
+     * The key id of the KMS. Used for encrypting a disk if not null. Only for PostgreSQL, MySQL.
      */
     public readonly encryptionKey!: pulumi.Output<string | undefined>;
     /**
@@ -401,7 +401,7 @@ export interface InstanceState {
      */
     readonly dbInstanceStorageType?: pulumi.Input<string>;
     /**
-     * The key id of the KMS. Used for encrypting a disk if not null. Only for PostgreSQL.
+     * The key id of the KMS. Used for encrypting a disk if not null. Only for PostgreSQL, MySQL.
      */
     readonly encryptionKey?: pulumi.Input<string>;
     /**
@@ -554,7 +554,7 @@ export interface InstanceArgs {
      */
     readonly dbInstanceStorageType?: pulumi.Input<string>;
     /**
-     * The key id of the KMS. Used for encrypting a disk if not null. Only for PostgreSQL.
+     * The key id of the KMS. Used for encrypting a disk if not null. Only for PostgreSQL, MySQL.
      */
     readonly encryptionKey?: pulumi.Input<string>;
     /**