+++ sdk/python/pulumi_alicloud/privatelink/__init__.py
@@ -3,9 +3,15 @@
 # *** Do not edit by hand unless you're certain you know what you are doing! ***
 
 # Export this package's modules as members:
+from .get_vpc_endpoint_connections import *
+from .get_vpc_endpoint_service_resources import *
+from .get_vpc_endpoint_service_users import *
 from .get_vpc_endpoint_services import *
 from .get_vpc_endpoints import *
 from .vpc_endpoint import *
 from .vpc_endpoint_service import *
+from .vpc_endpoint_service_connection import *
+from .vpc_endpoint_service_resource import *
+from .vpc_endpoint_service_user import *
 from ._inputs import *
 from . import outputs