+++ sdk/python/pulumi_alicloud/privatelink/_inputs.py
@@ -9,48 +9,8 @@ from typing import Any, Mapping, Optional, Sequence, Union
 from .. import _utilities, _tables
 
 __all__ = [
-    'VpcEndpointServiceResourceArgs',
     'VpcEndpointZoneArgs',
 ]
-
-@pulumi.input_type
-class VpcEndpointServiceResourceArgs:
-    def __init__(__self__, *,
-                 resource_id: Optional[pulumi.Input[str]] = None,
-                 resource_type: Optional[pulumi.Input[str]] = None):
-        """
-        :param pulumi.Input[str] resource_id: The id of service resources added to the endpoint service.
-        :param pulumi.Input[str] resource_type: The type of service resource added to the endpoint service.
-        """
-        if resource_id is not None:
-            pulumi.set(__self__, "resource_id", resource_id)
-        if resource_type is not None:
-            pulumi.set(__self__, "resource_type", resource_type)
-
-    @property
-    @pulumi.getter(name="resourceId")
-    def resource_id(self) -> Optional[pulumi.Input[str]]:
-        """
-        The id of service resources added to the endpoint service.
-        """
-        return pulumi.get(self, "resource_id")
-
-    @resource_id.setter
-    def resource_id(self, value: Optional[pulumi.Input[str]]):
-        pulumi.set(self, "resource_id", value)
-
-    @property
-    @pulumi.getter(name="resourceType")
-    def resource_type(self) -> Optional[pulumi.Input[str]]:
-        """
-        The type of service resource added to the endpoint service.
-        """
-        return pulumi.get(self, "resource_type")
-
-    @resource_type.setter
-    def resource_type(self, value: Optional[pulumi.Input[str]]):
-        pulumi.set(self, "resource_type", value)
-
 
 @pulumi.input_type
 class VpcEndpointZoneArgs: